% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExp2ddCt.R
\name{build_exp_plot}
\alias{build_exp_plot}
\title{Build expression plot (box or bar)}
\usage{
build_exp_plot(df.plot, fig.type, fig.ncol)
}
\description{
Build expression plot (box or bar)
}
\keyword{internal}
