% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalCurve.R
\name{cal_curve_by_mean}
\alias{cal_curve_by_mean}
\title{Build standard curve model using mean Cq values}
\usage{
cal_curve_by_mean(df, dilution)
}
\description{
Build standard curve model using mean Cq values
}
\keyword{internal}
