% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalCurve.R
\name{cal_curve_by_raw}
\alias{cal_curve_by_raw}
\title{Build standard curve model using raw Cq values}
\usage{
cal_curve_by_raw(df, dilution)
}
\description{
Build standard curve model using raw Cq values
}
\keyword{internal}
