% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExpCurve.R
\name{cal_curve_stat_test}
\alias{cal_curve_stat_test}
\title{Run statistical tests for curve-based expression}
\usage{
cal_curve_stat_test(df, stat.method, ref.group)
}
\description{
Run statistical tests for curve-based expression
}
\keyword{internal}
