% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-ddct.R
\name{calculate_ddct}
\alias{calculate_ddct}
\title{Calculate ddCT from dCT data.}
\usage{
calculate_ddct(dCT, factors, reference, metadata = NULL, genes = NULL)
}
\arguments{
\item{dCT}{Imported dCT qPCR data as a \code{data.table} (i.e., output of \code{read_dct()}).}

\item{factors}{A character vector listing factors for statistical comparison corresponding to a subset of the columns of each \code{data.table} object.}

\item{reference}{A character vector of the same length as factors, listing the level of each factor corresponding to the reference group.}

\item{metadata}{Columns in the dCT \code{data.table} which are not genes, but are not relevant for comparison. Either specify \code{metadata} or \code{genes}.}

\item{genes}{A character vector listing genes corresponding to a subset of the columns of each \code{data.table} object.}
}
\value{
ddCT data as a \code{data.table}.
}
\description{
Calculate delta delta CT from delta CT.
}
\author{
Ayush Noori
}
