% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExp2ddCt.R
\name{CalExp2ddCt}
\alias{CalExp2ddCt}
\title{Calculate Expression Using 2-ddCt Method}
\usage{
CalExp2ddCt(
  cq.table,
  design.table,
  ref.gene = "OsUBQ",
  ref.group = "CK",
  stat.method = "t.test",
  remove.outliers = TRUE,
  fig.type = "box",
  fig.ncol = NULL
)
}
\arguments{
\item{cq.table}{A data frame containing position and Cq values.
Must have columns: Position, Gene, Cq.}

\item{design.table}{A data frame containing position and group information.
Must have columns: Position, Group, BioRep.}

\item{ref.gene}{Character. The name of the reference gene (default: "OsUBQ").}

\item{ref.group}{Character. The name of the reference/control group
(default: "CK").}

\item{stat.method}{Character. Statistical method for group comparison.
One of "t.test", "wilcox.test", or "anova" (default: "t.test").}

\item{remove.outliers}{Logical. Remove outliers using IQR method
(default: TRUE).}

\item{fig.type}{Character. Plot type: "box" for boxplot, "bar" for barplot
(default: "box").}

\item{fig.ncol}{Integer. Number of columns in facet plot (default: NULL).}
}
\value{
A list containing:
  \item{table}{Data frame with expression values and statistics}
  \item{figure}{ggplot object}
}
\description{
Calculate relative gene expression using the 2-ddCt method
with a reference gene and reference group for normalization.
Supports statistical testing and outlier removal.
}
\examples{
\dontrun{
df1.path <- system.file("examples", "ddct.cq.txt", package = "qPCRtools")
df2.path <- system.file("examples", "ddct.design.txt", package = "qPCRtools")
cq.table <- read.table(df1.path, header = TRUE)
design.table <- read.table(df2.path, header = TRUE)
res <- CalExp2ddCt(
  cq.table,
  design.table,
  ref.gene = "OsUBQ",
  ref.group = "CK",
  stat.method = "t.test",
  remove.outliers = TRUE,
  fig.type = "box",
  fig.ncol = NULL
)
res[["table"]]
res[["figure"]]
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
