% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-variables.R
\name{define_variables}
\alias{define_variables}
\title{Define variables for statistical analysis.}
\usage{
define_variables(dCT, factors = NULL)
}
\arguments{
\item{dCT}{Imported dCT qPCR data (i.e., output of \code{read_dct()}).}

\item{format}{Format of the data file to import. Either \code{"csv"} to specify comma-separated data,}
}
\value{
Releveled dCT data.
}
\description{
Define variables and genes of dCT dataset in qPCR analysis
}
\author{
Ayush Noori
}
