% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPI.R
\name{easyPI}
\alias{easyPI}
\title{Easy percent input analysis}
\usage{
easyPI(
  infile = NULL,
  title = NULL,
  dilution = NULL,
  delim = NULL,
  input = NULL,
  refCond = NULL,
  method = "PI",
  std = NULL,
  avg = NULL,
  writeOut = NULL,
  col = "Dark2",
  returnDat = NULL,
  showStat = T,
  showEB = T
)
}
\arguments{
\item{infile}{Character indicating input file (raw from qPCR maching). Leave NULL to choose interactively.}

\item{title}{Character identifying the title of the plot. Leave NULL to choose interactively.}

\item{dilution}{Numeric indicating the diltion of the input as a percentage. e.g. 1:10 = 10 (10 percent)}

\item{delim}{Character indicating the delimiter in the 'Sample' column to separate conditions (entered first) and replicates (entered second). Leave NULL to choose interactively.}

\item{input}{Character indicating the input condition}

\item{refCond}{Character indicating the control condition to use for normalization. Leave NULL to choose interactively.}

\item{method}{Character indicating analysis method. Currently only percent input method available.}

\item{std}{Numeric indicating standard deviation threshold before evaluating outliers. Leave NULL to choose interactively.}

\item{avg}{Character indicating the averaging method ('geoMean', or 'mean'). Leave NULL to choose interactively.}

\item{writeOut}{Character ("Y"/"N") indicating if results are to be written out to csv. Leave NULL to choose interactively.}

\item{col}{Character indicating the RColorBrewer palette for colour scheme. Default is "Dark2".}

\item{returnDat}{Character ("Y"/"N") indicating if results should be returned to console. Leave NULL to choose interactively.}

\item{showStat}{Boolean indicating if statistics are to be shown. Default is TRUE.}

\item{showEB}{Boolean indicating if error bars are to be showns. Default is TRUE.}
}
\value{
data.frame object with relative expression values and a bar plot depicting relative expression
}
\description{
This function will make percent input anlaysis easy by asking for step-by-step
information, including input file, reference genes, and reference conditions.
}
