% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRT.R
\name{easyRT}
\alias{easyRT}
\title{Easy RT-qPCR analysis}
\usage{
easyRT(
  infile = NULL,
  title = NULL,
  refGene = NULL,
  delim = NULL,
  refCond = NULL,
  method = "ddCt",
  std = NULL,
  avg = NULL,
  writeOut = NULL,
  col = "Dark2",
  showStat = T,
  showEB = T
)
}
\arguments{
\item{infile}{Character indicating input file (raw from qPCR maching). Leave NULL to choose interactively.}

\item{title}{Character identifying the title of the plot. Leave NULL to choose interactively.}

\item{refGene}{Character vector indicating gene(s) to use for reference expression. Leave NULL to choose interactively.}

\item{delim}{Character indicating the delimiter in the 'Sample' column to separate conditions (entered first) and replicates (entered second). Leave NULL to choose interactively.}

\item{refCond}{Character indicating the control condition to use for normalization. Leave NULL to choose interactively.}

\item{method}{Character indicating analysis method. Currently only delta delta Ct method available.}

\item{std}{Numeric indicating standard deviation threshold before evaluating outliers. Leave NULL to choose interactively.}

\item{avg}{Character indicating the averaging method ('geoMean', or 'mean'). Leave NULL to choose interactively.}

\item{writeOut}{Character ("Y"/"N") indicating if results are to be written out to csv. Leave NULL to choose interactively.}

\item{col}{Character indicating the RColorBrewer palette for colour scheme. Default is "Dark2".}

\item{showStat}{Boolean indicating if statistics are to be shown. Default is TRUE.}

\item{showEB}{Boolean indicating if error bars are to be showns. Default is TRUE.}
}
\value{
data.frame object with relative expression values and a bar plot depicting relative expression
}
\description{
This function will make RT-qPCR anlaysis easy by asking for step-by-step
information, including input file, reference genes, and reference conditions.
}
