% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExpRqPCR.R
\name{find_ref_gene}
\alias{find_ref_gene}
\title{Find reference genes using GeNorm algorithm}
\usage{
find_ref_gene(df.expre)
}
\description{
Find reference genes using GeNorm algorithm
}
\keyword{internal}
