% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTUI.R
\name{PIUI}
\alias{PIUI}
\title{User Interface Shiny App for easyPI percent input function}
\usage{
PIUI()
}
\value{
A Shiny app that makes percent input analysis easier for ChIP-qPCR data
}
\description{
User Interface Shiny App for easyPI percent input function
}
