% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr-class.R
\name{qPCR}
\alias{qPCR}
\title{An S4 class to contain qPCR data.}
\usage{
qPCR(dCT, factors, reference, metadata = NULL, genes = NULL)
}
\arguments{
\item{dCT}{A \code{data.table} containing dCT data.}

\item{factors}{A character vector listing factors for statistical comparison corresponding to a subset of the columns of each \code{data.table} object.}

\item{reference}{A character vector of the same length as factors, listing the level of each factor corresponding to the reference group.}

\item{metadata}{Columns in the dCT \code{data.table} which are not genes, but are not relevant for comparison.}

\item{genes}{A character vector listing genes corresponding to a subset of the columns of each \code{data.table} object.}
}
\description{
The qPCR class is the main class of the \code{qpcrtools} package and is designed to contain all relevant qPCR data objects.
}
\examples{
\dontrun{qPCR(factors = c("Genotype", "Treatment"), reference = c("APOE3", "PBS"), metadata = c("Mouse", "Sex"))}
}
\author{
Ayush Noori
}
