% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrtools-package.R
\docType{package}
\name{qpcrtools-package}
\alias{qpcrtools}
\alias{qpcrtools-package}
\title{qpcrtools: Analysis and Visualization of RT-qPCR Data}
\description{
Tools for the analysis and visualization of real-time quantitative polymerase chain reaction (RT-qPCR) data in R.
}
\author{
Ayush Noori, Zhaozhi Li, Colin Magdamo, Sudeshna Das
}
\keyword{internal}
