% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dct.R
\name{read_dct}
\alias{read_dct}
\title{Read dCT data.}
\usage{
read_dct(file, format = "csv", ...)
}
\arguments{
\item{file}{File name of dCT data in working directory, or, alternatively, path to the qPCR data file.}

\item{format}{Format of the data file to import. Either \code{"csv"} to specify comma-separated data,
or the name or index of the sheet in the Excel workbook to read the data from.}

\item{...}{Additional arguments to be passed to \code{data.table::fread()} or \code{openxlsx::read.xlsx()} as needed.}
}
\value{
qPCR S4 object containing dCT data to be passed to next functions.
}
\description{
Read dCT data from qPCR experiment.
}
\author{
Ayush Noori
}
