% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTUI.R
\name{RTshiny}
\alias{RTshiny}
\title{easyRT for Shiny App}
\usage{
RTshiny(
  infile = NULL,
  title = NULL,
  refGene = NULL,
  delim = NULL,
  refCond = NULL,
  method = "ddCt",
  std = NULL,
  avg = NULL,
  writeOut = NULL,
  col = "Dark2",
  showStat = T,
  showEB = T,
  bioRad = NULL
)
}
\description{
easyRT for Shiny App
}
