% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTUI.R
\name{RTUI}
\alias{RTUI}
\title{User Interface Shiny App for easyRT delta-delta Ct function}
\usage{
RTUI()
}
\value{
A Shiny app that makes ddCt analysis easier for RT-qPCR data
}
\description{
User Interface Shiny App for easyRT delta-delta Ct function
}
