% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr-class.R
\name{show,qPCR-method}
\alias{show,qPCR-method}
\title{Display S4 object containing qPCR data.}
\usage{
\S4method{show}{qPCR}(object)
}
\arguments{
\item{object}{An S4 object of class \code{qPCR} containing qPCR data.}
}
\description{
Display S4 object containing qPCR data.
}
\author{
Ayush Noori
}
