% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExpRqPCR.R
\name{cal_rqpcr_stat_test}
\alias{cal_rqpcr_stat_test}
\title{Run statistical tests for RqPCR expression}
\usage{
cal_rqpcr_stat_test(res.all, stat.method, ref.group)
}
\description{
Run statistical tests for RqPCR expression
}
\keyword{internal}
