% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExp2ddCt.R
\name{cal_stat_test}
\alias{cal_stat_test}
\title{Run statistical tests on expression data}
\usage{
cal_stat_test(res.all, stat.method, ref.group)
}
\description{
Run statistical tests on expression data
}
\keyword{internal}
