% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalCurve.R
\name{CalCurve}
\alias{CalCurve}
\title{Standard Curve Calculation}
\usage{
CalCurve(
  cq_table,
  concen_table,
  highest_concen,
  lowest_concen,
  dilution = 4,
  by_mean = TRUE
)
}
\arguments{
\item{cq_table}{A data frame containing position and Cq values.
Must have columns: Position, Gene, Cq.}

\item{concen_table}{A data frame containing position and concentration.
Must have columns: Position, Conc.}

\item{highest_concen}{Numeric. The highest concentration for calculation.}

\item{lowest_concen}{Numeric. The lowest concentration for calculation.}

\item{dilution}{Numeric. Dilution factor of cDNA template (default: 4).}

\item{by_mean}{Logical. Calculate by mean Cq value or not (default: TRUE).}
}
\value{
A list containing:
  \item{table}{Data frame with standard curve parameters per gene
    (Formula, Slope, Intercept, R2, P.value, max.Cq, min.Cq, E, Date)}
  \item{figure}{ggplot object of the standard curve}
}
\description{
Calculate the standard curve and obtain the amplification efficiency
of primer(s). Based on the amplification efficiency, we can determine
which method to use for expression level calculation.
}
\examples{
\dontrun{
df.1.path <- system.file("examples", "calsc.cq.txt", package = "qPCRtools")
df.2.path <- system.file("examples", "calsc.info.txt", package = "qPCRtools")
df.1 <- read.table(df.1.path, header = TRUE)
df.2 <- read.table(df.2.path, header = TRUE)
res <- CalCurve(
  cq_table = df.1,
  concen_table = df.2,
  lowest_concen = 4,
  highest_concen = 4096,
  dilution = 4,
  by_mean = TRUE
)
res[["table"]]
res[["figure"]]
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
