% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExp2dCt.R
\name{CalExp2dCt}
\alias{CalExp2dCt}
\title{Calculate Expression Using 2-dCt Method}
\usage{
CalExp2dCt(cq_table, design_table, ref_gene = "Actin")
}
\arguments{
\item{cq_table}{A data frame containing position and Cq values.
Must have columns: Position, Gene, Cq.}

\item{design_table}{A data frame containing position and group information.
Must have columns: Position, Group, BioRep.}

\item{ref_gene}{Character. The name of the reference gene (default: "Actin").}
}
\value{
A data frame with expression values, including columns:
  position, cq, group, gene, biorep, mean.cq, expre, n, mean.expre,
  sd.expre, se.expre.
}
\description{
Calculate relative gene expression using the 2-dCt method
with a reference gene for normalization.
}
\examples{
\dontrun{
df1.path <- system.file("examples", "dct.cq.txt", package = "qPCRtools")
df2.path <- system.file("examples", "dct.design.txt", package = "qPCRtools")
cq_table <- read.table(df1.path, sep = ",", header = TRUE)
design_table <- read.table(df2.path, sep = ",", header = TRUE)
res <- CalExp2dCt(cq_table, design_table, ref_gene = "Actin")
head(res)
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
