% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalExpRqPCR.R
\name{CalExpRqPCR}
\alias{CalExpRqPCR}
\title{Calculate Expression Using RqPCR Method}
\usage{
CalExpRqPCR(
  cq.table,
  design.table,
  ref.gene = NULL,
  ref.group = "CK",
  stat.method = "t.test",
  fig.type = "box",
  fig.ncol = NULL
)
}
\arguments{
\item{cq.table}{A data frame containing position and Cq values.
Must have columns: Position, Gene, Cq, BioRep, TechRep, Eff.}

\item{design.table}{A data frame containing position and group information.
Must have columns: Position, Group, BioRep, TechRep, Eff.}

\item{ref.gene}{Character. The name(s) of reference gene(s).
If NULL, reference genes are auto-selected via GeNorm (default: NULL).}

\item{ref.group}{Character. The name of the reference/control group
(default: "CK").}

\item{stat.method}{Character. Statistical method for group comparison.
One of "t.test", "wilcox.test", or "anova" (default: "t.test").}

\item{fig.type}{Character. Plot type: "box" for boxplot, "bar" for barplot
(default: "box").}

\item{fig.ncol}{Integer. Number of columns in facet plot (default: NULL).}
}
\value{
A list containing:
  \item{table}{Data frame with expression values and statistics}
  \item{figure}{ggplot object}
}
\description{
Calculate relative gene expression using the RqPCR method with
amplification efficiency correction. Can auto-select reference genes
using the GeNorm algorithm when ref.gene is NULL.
}
\examples{
\dontrun{
df1.path <- system.file("examples", "cal.expre.rqpcr.cq.txt", package = "qPCRtools")
df2.path <- system.file("examples", "cal.expre.rqpcr.design.txt", package = "qPCRtools")
cq.table <- read.table(df1.path, header = TRUE)
design.table <- read.table(df2.path, header = TRUE)
res <- CalExpRqPCR(
  cq.table, design.table,
  ref.gene = NULL,
  ref.group = "CK",
  stat.method = "t.test",
  fig.type = "box",
  fig.ncol = NULL
)
res[["table"]]
res[["figure"]]
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
