% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalRTable.R
\name{CalRTable}
\alias{CalRTable}
\title{Calculate RNA Volume for Reverse Transcription}
\usage{
CalRTable(data, template, RNA.weight = 1)
}
\arguments{
\item{data}{A data frame containing sample names and concentration
values (default unit: ng/uL). Must have columns: sample, concentration.}

\item{template}{A data frame containing reverse transcription information.
Must have a column called `all`.}

\item{RNA.weight}{Numeric. RNA weight required for reverse transcription
in micrograms (default: 1).}
}
\value{
A data frame with calculated RNA and water volumes for each sample.
}
\description{
The first step of qPCR is usually the preparation of cDNA.
This function calculates the volume of RNA needed for reverse
transcription based on RNA concentration.
}
\examples{
\dontrun{
df.1.path <- system.file("examples", "crtv.data.txt", package = "qPCRtools")
df.2.path <- system.file("examples", "crtv.template.txt", package = "qPCRtools")
df.1 <- read.table(df.1.path, sep = "\t", header = TRUE)
df.2 <- read.table(df.2.path, sep = "\t", header = TRUE)
result <- CalRTable(data = df.1, template = df.2, RNA.weight = 2)
head(result)
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
