% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr-class.R
\name{qPCR}
\alias{qPCR}
\title{An S4 class to contain qPCR data.}
\usage{
qPCR(
  dCT = data.table(),
  ddCT = data.table(),
  RQ = data.table(),
  genes = colnames(dCT),
  factors = character()
)

qPCR(
  dCT = data.table(),
  ddCT = data.table(),
  RQ = data.table(),
  genes = colnames(dCT),
  factors = character()
)
}
\arguments{
\item{dCT}{A \code{data.table} containing dCT data.}

\item{ddCT}{A \code{data.table} containing ddCT data.}

\item{RQ}{A \code{data.table} containing RQ data.}

\item{genes}{A character vector listing genes corresponding to a subset of the columns of each \code{data.table} object.}

\item{factors}{A character vector listing factors for statistical comparison corresponding to a subset of the columns of each \code{data.table} object.}
}
\description{
An S4 class to contain qPCR data.

An S4 class to contain qPCR data.
}
\section{Slots}{

\describe{
\item{\code{dCT}}{A \code{data.table} containing dCT data.}

\item{\code{ddCT}}{A \code{data.table} containing ddCT data.}

\item{\code{RQ}}{A \code{data.table} containing RQ data.}

\item{\code{genes}}{A character vector listing genes corresponding to a subset of the columns of each \code{data.table} object.}

\item{\code{factors}}{A character vector listing factors for statistical comparison corresponding to a subset of the columns of each \code{data.table} object.}
}}

\author{
Ayush Noori

Ayush Noori
}
